requirements.txt
selenium
pytest
python-dotenv
pytest-xdist
pytest-rerunfailures
pytest-html
